[FE] refactor: Dashboard
해당 월 지출 내역으로만 작동 될수 있게끔 수정
복잡한 로직 분리
결제 방식별 지출의 총 지출을 지출만 표시 되게끔 수정
